fix: Update InstanceV1/V2 related types based on real-world usage (#476)
Many servers that claim to be Mastodon-API compatible are not, as
evidenced by the content they include in the responses to
`/api/v1/instance` and `/api/v2/instance` requests.

Work around the worst of the breakage by providing defaults or marking
some fields as nullable (with a default null).

Bugs have been reported against the relevant projects.

+++ app/src/main/java/app/pachli/components/compose/MediaUploader.kt
@@ -93,7 +93,7 @@ fun createNewImageFile(context: Context, suffix: String = ".jpg"): File {
 
 data class PreparedMedia(val type: QueuedMedia.Type, val uri: Uri, val size: Long)
 
-class FileSizeException(val allowedSizeInBytes: Int) : Exception()
+class FileSizeException(val allowedSizeInBytes: Long) : Exception()
 class MediaTypeException : Exception()
 class CouldNotOpenFileException : Exception()
 class UploadServerError(val errorMessage: String) : Exception()
